requirements.txt
playwright==1.40.0
pytest-playwright==0.4.3
pytest-reporter-html1==0.8.3
pytest-xdist==3.4.0
pytest-dotenv==0.5.2
Faker==20.0.0
allure-pytest==2.13.2
pydantic==2.5.1